requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0

# Email and text processing
beautifulsoup4==4.12.2
# textstat==0.7.3  # Removed due to pkg_resources deprecation
email-validator==2.0.0

# Production dependencies
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4

# Security
cryptography==41.0.7
bcrypt==4.1.2

# Monitoring and logging
sentry-sdk[flask]==1.38.0
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Fix for pkg_resources issue
setuptools>=65.0.0 